update bulk unassign code to pass new tests

+++ lib/policy_machine_storage_adapters/active_record.rb
@@ -193,8 +193,10 @@ module PolicyMachineStorageAdapter
       end
 
       def self.bulk_unassign(pairs_hash)
-        pairs_str = pairs_hash.values.map { |parent,child| "(#{parent.id},#{child.id})" }.join(",")
-        Assignment.where("(parent_id,child_id) IN (#{pairs_str})").delete_all
+        pairs_str = pairs_hash.values.reduce([]) do |memo,(parent,child)|
+          parent.persisted && child.persisted? ? memo + ["(#{parent.id},#{child.id})"] : memo
+        end.join(',')
+        Assignment.where("(parent_id,child_id) IN (#{pairs_str})").delete_all unless pairs_str.empty?
       end
 
       def self.bulk_link(parents_and_children)
